# Conflicts:
#	BookYourSport/Services/Reservation/ReservationService.API/appsettings.json
#	BookYourSport/Services/Reservation/docker-compose.yml

+++ BookYourSport/Payment/PaymentService.Application/Commands/ChargeCredit/ChargeCreditCommand.cs
@@ -2,5 +2,5 @@
 
 public record ChargeCreditCommand(
     Guid UserId,
-    int Amount,
+    decimal Amount,
     Guid ReferenceId);
